refactor(console): extract api details quality to dedicated component
