+++ gravitee-apim-console-webui/src/management/api/details/api-portal-details-quality/api-portal-details-quality.component.html
@@ -1,0 +1,38 @@
+<!--
+
+    Copyright (C) 2015 The Gravitee team (http://gravitee.io)
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+            http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+-->
+<mat-card class="api-quality">
+  <h3>Quality</h3>
+  <div *ngIf="qualityMetrics" class="api-quality__content">
+    <gio-circular-percentage class="api-quality__content__score" [score]="qualityMetrics.score"></gio-circular-percentage>
+    <div class="api-quality__content__list">
+      <div *ngFor="let qualityMetric of qualityMetrics.metrics_passed | keyvalue" class="api-quality__content__list__item">
+        <mat-icon
+          *ngIf="qualityMetric.value"
+          class="api-quality__content__list__item__green-icon"
+          svgIcon="gio:check-circled-outline"
+        ></mat-icon>
+        <mat-icon
+          *ngIf="!qualityMetric.value"
+          class="api-quality__content__list__item__red-icon"
+          svgIcon="gio:warning-circled-outline"
+        ></mat-icon>
+        <span class="api-quality__content__list__item__text">{{ qualityMetricsDescription[qualityMetric.key] }}</span>
+      </div>
+    </div>
+  </div>
+</mat-card>
