+++ gravitee-apim-console-webui/src/management/api/details/api-portal-details.component.scss
@@ -113,43 +113,6 @@ $typography: map.get(gio.$mat-theme, typography);
 
 .api-quality {
   margin-top: 16px;
-
-  &__content {
-    display: flex;
-    align-items: center;
-    flex-direction: row;
-
-    &__score {
-      flex: 1 1 100%;
-      margin: 32px;
-    }
-
-    &__list {
-      flex: 1 1 auto;
-      display: flex;
-      align-items: center;
-      flex-wrap: wrap;
-      padding: 4px 0;
-
-      &__item {
-        width: 50%;
-        display: flex;
-        align-items: center;
-
-        &__text {
-          margin-left: 8px;
-        }
-
-        &__red-icon {
-          color: mat.get-color-from-palette(gio.$mat-error-palette, 'default');
-        }
-
-        &__green-icon {
-          color: mat.get-color-from-palette(gio.$mat-success-palette, 'default');
-        }
-      }
-    }
-  }
 }
 
 .danger-card {
