+++ gravitee-apim-console-webui/src/management/api/details/api-portal-details.module.ts
@@ -29,6 +29,7 @@ import { MatDialogModule } from '@angular/material/dialog';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 
 import { ApiPortalDetailsComponent } from './api-portal-details.component';
+import { ApiPortalDetailsQualityComponent } from './api-portal-details-quality/api-portal-details-quality.component';
 
 import { GioAvatarModule } from '../../../shared/components/gio-avatar/gio-avatar.module';
 import { GioFormFocusInvalidModule } from '../../../shared/components/gio-form-focus-first-invalid/gio-form-focus-first-invalid.module';
@@ -38,7 +39,7 @@ import { GioFormSlideToggleModule } from '../../../shared/components/gio-form-sl
 import { GioCircularPercentageModule } from '../../../shared/components/gio-circular-percentage/gio-circular-percentage.module';
 
 @NgModule({
-  declarations: [ApiPortalDetailsComponent],
+  declarations: [ApiPortalDetailsComponent, ApiPortalDetailsQualityComponent],
   exports: [ApiPortalDetailsComponent],
   imports: [
     CommonModule,
